# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          }
        },
        "summary": "Criar um novo usuário",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso."
          },
          "201": {
            "description": ""
          },
          "401": {
            "description": "Credenciais inválidas."
          }
        },
        "summary": "Fazer login",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Informações do usuário retornadas com sucesso."
          },
          "401": {
            "description": "Não autorizado."
          },
          "403": {
            "description": "Acesso negado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter informações do usuário autenticado",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/validate-token": {
      "post": {
        "operationId": "AuthController_validateToken",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token válido."
          },
          "201": {
            "description": ""
          },
          "401": {
            "description": "Token inválido ou expirado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Validar token de acesso",
        "tags": [
          "auth"
        ]
      }
    },
    "/funcionario": {
      "post": {
        "operationId": "FuncionarioController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFuncionarioDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Funcionário criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          }
        },
        "summary": "Criar um novo funcionário",
        "tags": [
          "funcionario"
        ]
      },
      "get": {
        "operationId": "FuncionarioController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de funcionários retornada com sucesso."
          }
        },
        "summary": "Listar todos os funcionários",
        "tags": [
          "funcionario"
        ]
      }
    },
    "/funcionario/{id}": {
      "get": {
        "operationId": "FuncionarioController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do funcionário",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Funcionário encontrado."
          },
          "404": {
            "description": "Funcionário não encontrado."
          }
        },
        "summary": "Obter um funcionário pelo ID",
        "tags": [
          "funcionario"
        ]
      },
      "put": {
        "operationId": "FuncionarioController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do funcionário",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFuncionarioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Funcionário atualizado com sucesso."
          },
          "404": {
            "description": "Funcionário não encontrado."
          }
        },
        "summary": "Atualizar um funcionário pelo ID",
        "tags": [
          "funcionario"
        ]
      },
      "delete": {
        "operationId": "FuncionarioController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do funcionário",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Funcionário removido com sucesso."
          },
          "404": {
            "description": "Funcionário não encontrado."
          }
        },
        "summary": "Remover um funcionário pelo ID",
        "tags": [
          "funcionario"
        ]
      }
    },
    "/gestao": {
      "post": {
        "operationId": "GestaoController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGestorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Gestor criado com sucesso."
          },
          "403": {
            "description": "Acesso negado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar um novo gestor",
        "tags": [
          "gestao"
        ]
      },
      "get": {
        "operationId": "GestaoController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de gestores retornada com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todos os gestores",
        "tags": [
          "gestao"
        ]
      }
    },
    "/gestao/{id}": {
      "get": {
        "operationId": "GestaoController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do gestor",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Gestor encontrado."
          },
          "404": {
            "description": "Gestor não encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter um gestor pelo ID",
        "tags": [
          "gestao"
        ]
      },
      "put": {
        "operationId": "GestaoController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do gestor",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGestorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Gestor atualizado com sucesso."
          },
          "404": {
            "description": "Gestor não encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar um gestor pelo ID",
        "tags": [
          "gestao"
        ]
      },
      "delete": {
        "operationId": "GestaoController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do gestor",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Gestor removido com sucesso."
          },
          "404": {
            "description": "Gestor não encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remover um gestor pelo ID",
        "tags": [
          "gestao"
        ]
      }
    },
...
  },
  "info": {
    "title": "NestJS API",
    "description": "API documentation for NestJS with Swagger, Orval, and MikroORM",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateAuthDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Código do usuário",
            "example": "FUNC123"
          },
          "password": {
            "type": "string",
            "description": "Senha do usuário",
            "example": "senha123"
          },
          "role": {
            "type": "string",
            "description": "Papel do usuário",
            "example": "gestao",
            "enum": [
              "funcionario",
              "gestao"
            ]
          }
        },
        "required": [
          "code",
          "password",
          "role"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Código do usuário",
            "example": "FUNC123"
          },
          "password": {
            "type": "string",
            "description": "Senha do usuário",
            "example": "senha123"
          }
        },
        "required": [
          "code",
          "password"
        ]
      },
      "CreateFuncionarioDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Código do funcionário"
          },
          "nome": {
            "type": "string",
            "description": "Nome do funcionário"
          },
          "cargo": {
            "type": "string",
            "description": "Cargo do funcionário"
          },
          "salario": {
            "type": "number",
            "description": "Salário do funcionário"
          },
          "password": {
            "type": "string",
            "description": "Senha para acesso do funcionário",
            "minLength": 6
          }
        },
        "required": [
          "code",
          "nome",
          "cargo",
          "salario",
          "password"
        ]
      },
      "UpdateFuncionarioDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Código do funcionário"
          },
          "nome": {
            "type": "string",
            "description": "Nome do funcionário"
          },
          "cargo": {
            "type": "string",
            "description": "Cargo do funcionário"
          },
          "salario": {
            "type": "number",
            "description": "Salário do funcionário"
          },
          "password": {
            "type": "string",
            "description": "Nova senha para acesso do funcionário (mínimo 6 caracteres)",
            "minLength": 6
          }
        }
      },
      "CreateGestorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Código do gestor"
          },
          "name": {
            "type": "string",
            "description": "Nome do gestor"
          },
          "department": {
            "type": "string",
            "description": "Departamento do gestor"
          },
          "role": {
            "type": "string",
            "description": "Cargo do gestor"
          },
          "password": {
            "type": "string",
            "description": "Senha para acesso do gestor",
            "minLength": 6
          }
        },
        "required": [
          "code",
          "name",
          "department",
          "role",
          "password"
        ]
      },
      "UpdateGestorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Código do gestor"
          },
          "name": {
            "type": "string",
            "description": "Nome do gestor"
          },
          "department": {
            "type": "string",
            "description": "Departamento do gestor"
          },
          "role": {
            "type": "string",
            "description": "Cargo do gestor"
          },
          "password": {
            "type": "string",
            "description": "Nova senha para acesso do gestor (mínimo 6 caracteres)",
            "minLength": 6
          }
        }
      },
...
    }
  }
}